fix(compiler): finish nodejs-next integration blockers

+++ packages/emitter/testcases/common/expected/edge-cases/void-expression/VoidExpression.cs
@@ -5,7 +5,7 @@ namespace TestCases.common.edgecases.voidexpression
     {
         public static void voidStatementMarker()
         {
-            int x = 1;
+            int x = (int)1;
             var __tsonic_discard = x;
         }
 
@@ -26,7 +26,7 @@ namespace TestCases.common.edgecases.voidexpression
 
         internal static int sideEffect()
         {
-            return 42;
+            return (int)42;
         }
     }
 }
