+++ packages/emitter/testcases/common/expected/extensions/system/Overlaps.cs
@@ -7,8 +7,8 @@ namespace TestCases.common.extensions.system
         {
             var s = "hello";
             var a = global::System.MemoryExtensions.AsSpan(((string)s));
-            var b = global::System.MemoryExtensions.AsSpan(((string)s), 1);
-            int off = 0;
+            var b = global::System.MemoryExtensions.AsSpan(((string)s), (int)1);
+            int off = (int)0;
             var ok1 = global::System.MemoryExtensions.Overlaps(a, b);
             var ok2 = global::System.MemoryExtensions.Overlaps(a, b, out off);
             global::System.Console.WriteLine($"ok1: {ok1}");
