+++ packages/emitter/testcases/common/expected/types/expected-type-threading/TernaryTyping.cs
@@ -3,13 +3,13 @@ namespace TestCases.common.types.expectedtypethreading
     [global::Tsonic.Internal.ModuleContainerAttribute]
     public static class TernaryTyping
     {
-        public static readonly int ternaryInt = true ? 5 : 10;
+        public static readonly int ternaryInt = true ? (int)5 : (int)10;
 
-        public static readonly int nestedTernary = true ? false ? 1 : 2 : 3;
+        public static readonly int nestedTernary = true ? false ? (int)1 : (int)2 : (int)3;
 
         public static int ternaryReturn(bool flag)
         {
-            return flag ? 100 : 200;
+            return flag ? (int)100 : (int)200;
         }
     }
 }
